double to float

+++ srcs_bonus/config_parse/map_validation.c
@@ -69,8 +69,8 @@ void	map_set_player(int pl_x, int pl_y, t_cub *cub)
 {
 	char dir;
 
-	cub->cam->step = (double)cub->map->blk_x / STEP_SCALER;
-	cub->cam->dst_to_wall = (double)cub->map->blk_x / WALL_SCALER;
+	cub->cam->step = (float)cub->map->blk_x / STEP_SCALER;
+	cub->cam->dst_to_wall = (float)cub->map->blk_x / WALL_SCALER;
 	cub->cam->x = cub->map->blk_x * pl_x + cub->map->blk_x / 2;
 	cub->cam->y = cub->map->blk_y * pl_y + cub->map->blk_y / 2;
 	dir = cub->map->field[pl_y][pl_x];
