+++ srcs_bonus/game/shadows.c
@@ -12,25 +12,25 @@
 
 #include "cub3d_bonus.h"
 
-unsigned int	add_shadow(unsigned int color, double len, t_cub *cub)
+unsigned int	add_shadow(unsigned int color, float len, t_cub *cub)
 {
-	static double	len_no_shadow;
-	static double	len_full_shadow;
-	double			off_shadow;
+	static float	len_no_shadow;
+	static float	len_full_shadow;
+	float			off_shadow;
 
 	if (!len_no_shadow)
-		len_no_shadow = (double)cub->map->blk_x * NO_SHADOW_DST;
+		len_no_shadow = (float)cub->map->blk_x * NO_SHADOW_DST;
 	if (!len_full_shadow)
-		len_full_shadow = (double)cub->map->blk_x * (FULL_SHADOW_DST - NO_SHADOW_DST);
+		len_full_shadow = (float)cub->map->blk_x * (FULL_SHADOW_DST - NO_SHADOW_DST);
 	if ((off_shadow = len - len_no_shadow) < 0)
 		return (color);
 	if (off_shadow > len_full_shadow)
 		return (0);
-	return (decrease_color(color, 1. - off_shadow / len_full_shadow));
+	return (decrease_color(color, 1.f - off_shadow / len_full_shadow));
 
 }
 
-unsigned int	decrease_color(unsigned int color, double scaler)
+unsigned int	decrease_color(unsigned int color, float scaler)
 {
 	unsigned int	new_color;
