+++ srcs_bonus/game/walls.c
@@ -12,26 +12,26 @@
 
 #include "cub3d_bonus.h"
 
-double			get_x_wall(t_cub *cub)
+float			get_x_wall(t_cub *cub)
 {
-	double not_used;
+	float not_used;
 
 	if (cub->ray->dir == DIR_RIGHT)
-		return (1. - modf(cub->ray->y / (double)cub->map->blk_y, &not_used));
+		return (1.f - modff(cub->ray->y / (float)cub->map->blk_y, &not_used));
 	else if (cub->ray->dir == DIR_BOT)
-		return (modf(cub->ray->x / (double)cub->map->blk_x, &not_used));
+		return (modff(cub->ray->x / (float)cub->map->blk_x, &not_used));
 	else if (cub->ray->dir == DIR_LEFT)
-		return (modf(cub->ray->y / (double)cub->map->blk_y, &not_used));
+		return (modff(cub->ray->y / (float)cub->map->blk_y, &not_used));
 	else
-		return (1. - modf(cub->ray->x / (double)cub->map->blk_x, &not_used));
+		return (1.f - modff(cub->ray->x / (float)cub->map->blk_x, &not_used));
 }
 
-double			get_y_wall(double angle, double len_to_wall, t_cub *cub)
+float			get_y_wall(float angle, float len_to_wall, t_cub *cub)
 {
-	return ((cub->cam->z + len_to_wall * tan(angle)) / cub->map->blk_z);
+	return ((cub->cam->z + len_to_wall * tanf(angle)) / cub->map->blk_z);
 }
 
-unsigned int	get_pixel_wall(double off_x, double off_y, t_cub *cub,
+unsigned int	get_pixel_wall(float off_x, float off_y, t_cub *cub,
 							   int is_sp)
 {
 	t_img *no;
@@ -40,23 +40,23 @@ unsigned int	get_pixel_wall(double off_x, double off_y, t_cub *cub,
 	t_img *we;
 
 	if (is_sp)
-		return ((int*)cub->tex->s0->data)[(int)((double)cub->tex->s0->height
+		return ((int*)cub->tex->s0->data)[(int)((float)cub->tex->s0->height
 		* off_y) * cub->tex->s0->width +
-		(int)((double)cub->tex->s0->width * off_x)];
+		(int)((float)cub->tex->s0->width * off_x)];
 	no = cub->tex->no;
 	so = cub->tex->so;
 	ea = cub->tex->ea;
 	we = cub->tex->we;
 	if (cub->ray->dir == DIR_TOP)
-		return ((int*)no->data)[(int)((double)no->height * off_y) * no->width
-		+ (int)((double)no->width * off_x)];
+		return ((int*)no->data)[(int)((float)no->height * off_y) * no->width
+		+ (int)((float)no->width * off_x)];
 	else if (cub->ray->dir == DIR_BOT)
-		return ((int*)so->data)[(int)((double)so->height * off_y) * so->width
-		+ (int)((double)so->width * off_x)];
+		return ((int*)so->data)[(int)((float)so->height * off_y) * so->width
+		+ (int)((float)so->width * off_x)];
 	else if (cub->ray->dir == DIR_LEFT)
-		return ((int*)ea->data)[(int)((double)ea->height * off_y) * ea->width
-		+ (int)((double)ea->width * off_x)];
+		return ((int*)ea->data)[(int)((float)ea->height * off_y) * ea->width
+		+ (int)((float)ea->width * off_x)];
 	else
-		return ((int*)we->data)[(int)((double)we->height * off_y) * we->width
-			+ (int)((double)we->width * off_x)];
+		return ((int*)we->data)[(int)((float)we->height * off_y) * we->width
+			+ (int)((float)we->width * off_x)];
 }
